Revert "feat: add contributor activity line (#23)"
This reverts commit 7f002b1eeca7bd112249ca1f78a42bda852c39e3.
