+++ src/App.js
@@ -1,18 +1,25 @@
 import React from "react";
-import { makeStyles, Paper, IconButton, Snackbar } from "@material-ui/core";
+import cloneDeep from "lodash.clonedeep";
+import omit from "lodash.omit";
+import { Row, Col, Tab } from "react-bootstrap";
+import ReactECharts from "echarts-for-react";
+
+import {
+  Button,
+  ButtonGroup,
+  makeStyles,
+  Paper,
+  IconButton,
+  Snackbar
+} from "@material-ui/core";
 import TextField from "@material-ui/core/TextField";
 import SearchIcon from "@material-ui/icons/Search";
 import MuiAlert from "@material-ui/lab/Alert";
 import Autocomplete from "@material-ui/lab/Autocomplete";
-import Tabs from "@material-ui/core/Tabs";
-import Tab from "@material-ui/core/Tab";
-import PropTypes from "prop-types";
-import Typography from "@material-ui/core/Typography";
-import Box from "@material-ui/core/Box";
 
-import ContirbutorLineChart from "./components/contributor";
-import ActivityChart from "./components/activity";
-import { getParameterByName } from "./utils";
+import CompareComponent from "./components/compare";
+import { getMonths, getParameterByName, isSameDay } from "./utils";
+import { DEFAULT_OPTIONS } from "./constants";
 
 const Alert = props => {
   return <MuiAlert elevation={6} variant="filled" {...props} />;
@@ -38,7 +45,7 @@ const useStyles = makeStyles(theme => ({
   },
   autocomplete: {
     marginLeft: theme.spacing(1),
-    flex: 1
+    flex: 1,
   },
   iconButton: {
     padding: 10
@@ -49,67 +56,17 @@ const useStyles = makeStyles(theme => ({
   }
 }));
 
-function TabPanel(props) {
-  const { children, value, index, ...other } = props;
-
-  return (
-    <div
-      role="tabpanel"
-      hidden={value !== index}
-      id={`vertical-tabpanel-${index}`}
-      aria-labelledby={`vertical-tab-${index}`}
-      {...other}
-      style={{ width: "100%" }}
-    >
-      {value === index && (
-        <Box p={6} style={{ padding: 0 }}>
-          <Typography>{children}</Typography>
-        </Box>
-      )}
-    </div>
-  );
-}
-
-TabPanel.propTypes = {
-  children: PropTypes.node,
-  index: PropTypes.any.isRequired,
-  value: PropTypes.any.isRequired
-};
-
-function a11yProps(index) {
-  return {
-    id: `vertical-tab-${index}`,
-    "aria-controls": `vertical-tabpanel-${index}`
-  };
-}
-
-const useStylesTable = makeStyles(theme => ({
-  root: {
-    flexGrow: 1,
-    backgroundColor: theme.palette.background.paper,
-    display: "flex",
-    height: 224,
-    marginTop: "2em"
-  },
-  tabs: {
-    borderRight: `1px solid ${theme.palette.divider}`,
-    // display: window.innerWidth > 555 ? "" : "none"
-  }
-}));
-
 const App = () => {
+  const [loading, setLoading] = React.useState(false);
+  const [dataSource, setDataSource] = React.useState({});
+  const [activeDate, setActiveDate] = React.useState("max");
+  const [xAxis, setXAxis] = React.useState([]);
+  const [option, setOption] = React.useState(DEFAULT_OPTIONS);
   const [repo, setRepo] = React.useState("apache/apisix");
   const [message, setMessage] = React.useState("");
   const [open, setOpen] = React.useState(false);
   const [alertType, setAlertType] = React.useState("success");
   const [searchOption, setSearchOption] = React.useState([]);
-  const [contributorRepoList, setContributorRepoList] = React.useState([]);
-  const classesTable = useStylesTable();
-  const [value, setValue] = React.useState(0);
-
-  const handleChange = (event, newValue) => {
-    setValue(newValue);
-  };
 
   const classes = useStyles();
 
@@ -127,10 +84,134 @@ const App = () => {
     setOpen(false);
   };
 
-  const updateChart = repo => {
-    if (!contributorRepoList.includes(repo)) {
-      setContributorRepoList([...contributorRepoList, repo]);
+  const updateSeries = passXAxis => {
+    const newClonedOption = cloneDeep(DEFAULT_OPTIONS);
+    const datasetWithFilters = [
+      ["ContributorNum", "Repo", "Date", "DateValue"]
+    ];
+    const legend = [];
+    const limitDate = new Date(passXAxis[0]).getTime();
+
+    Object.entries(dataSource).forEach(([key, value]) => {
+      legend.push(key);
+      value.forEach(item => {
+        datasetWithFilters.push([
+          item.contributorNum,
+          item.repo,
+          item.date,
+          new Date(item.date).getTime()
+        ]);
+      });
+    });
+
+    const newDateSet = datasetWithFilters.sort(
+      (a, b) => new Date(a[2]) - new Date(b[2])
+    );
+
+    const filterDataset = legend.map(item => ({
+      id: item,
+      fromDatasetId: "dataset_raw",
+      transform: {
+        type: "filter",
+        config: {
+          and: [
+            { dimension: "Repo", "=": item },
+            { dimension: "DateValue", gte: limitDate }
+          ]
+        }
+      }
+    }));
+
+    const series = legend.map(item => ({
+      name: item,
+      type: "line",
+      datasetId: item,
+      showSymbol: false,
+      encode: {
+        x: "Date",
+        y: "ContributorNum",
+        itemName: "Repo",
+        tooltip: ["ContributorNum"]
+      }
+    }));
+
+    newClonedOption.dataset = [
+      {
+        id: "dataset_raw",
+        source: newDateSet
+      }
+    ].concat(filterDataset);
+
+    newClonedOption.series = series;
+    newClonedOption.legend.data = legend;
+
+    setOption(newClonedOption);
+  };
+
+  const fetchData = repo => {
+    if (repo === "null" || repo === null) {
+      repo = "apache/apisix";
     }
+    setLoading(true);
+
+    return new Promise((resolve, reject) => {
+      fetch(
+        `https://contributor-graph-api.apiseven.com/contributors?repo=${repo}`
+      )
+        .then(response => {
+          if (!response.ok) {
+            let message = "";
+            switch (response.status) {
+              case 403:
+                message = "Hit rate limit";
+                break;
+              case 404:
+                message = "Repo format error / Repo not found";
+                break;
+              default:
+                message = "Request Error";
+                break;
+            }
+            throw message;
+          }
+          return response.json();
+        })
+        .then(myJson => {
+          setLoading(false);
+          resolve({ repo, ...myJson });
+        })
+        .catch(e => {
+          showAlert(e, "error");
+          setLoading(false);
+          reject();
+        });
+    });
+  };
+
+  const updateChart = repo => {
+    if (dataSource[repo]) return;
+
+    fetchData(repo).then(myJson => {
+      const { Contributors = [] } = myJson;
+      const data = Contributors.map(item => ({
+        repo,
+        contributorNum: item.idx,
+        date: item.date
+      }));
+
+      if (!isSameDay(new Date(data[data.length - 1].date), new Date())) {
+        data.push({
+          repo,
+          contributorNum: Contributors[Contributors.length - 1].idx,
+          date: new Date()
+        });
+      }
+
+      const clonedDatasource = cloneDeep(dataSource);
+      if (!clonedDatasource[repo]) {
+        setDataSource({ ...clonedDatasource, ...{ [repo]: data } });
+      }
+    });
   };
 
   const getSearchOptions = () => {
@@ -152,25 +233,58 @@ const App = () => {
   };
 
   React.useEffect(() => {
-    const repo = getParameterByName("repo");
-    const chart = getParameterByName("chart");
-    if (chart === "contributorMonthlyActivity") {
-      setValue(1);
+    switch (activeDate) {
+      case "1month":
+        setXAxis(getMonths(1));
+        break;
+      case "3months":
+        setXAxis(getMonths(3));
+        break;
+      case "6months":
+        setXAxis(getMonths(6));
+        break;
+      case "1year":
+        setXAxis(getMonths(12));
+        break;
+      case "max":
+        setXAxis(["1970-01-01"]);
+        break;
+      default:
+        break;
     }
+  }, [activeDate]);
+
+  React.useEffect(() => {
+    updateSeries(xAxis);
+    window.parent.postMessage(Object.keys(dataSource), "*");
+  }, [dataSource, xAxis]);
+
+  React.useEffect(() => {
+    const repo = getParameterByName("repo");
     if (repo) {
       const repoArr = repo.split(",").filter(Boolean);
-      setContributorRepoList(repoArr);
+      Promise.all(repoArr.map(item => fetchData(item))).then(data => {
+        const tmpDataSouce = {};
+        data.forEach(item => {
+          const { Contributors = [], repo } = item;
+          const data = Contributors.map(item => ({
+            repo,
+            contributorNum: item.idx,
+            date: item.date
+          }));
+
+          if (!tmpDataSouce[item.repo]) {
+            tmpDataSouce[repo] = data;
+          }
+        });
+        setDataSource(tmpDataSouce);
+      });
     } else {
-      setContributorRepoList(["apache/apisix"]);
+      updateChart("apache/apisix");
     }
+
     getSearchOptions();
   }, []);
-
-  React.useEffect(()=> {
-    window.parent.postMessage({
-      chartType: value === 0 ? 'contributorOverTime':'contributorMonthlyActivity'
-    }, "*");
-  },[value])
 
   return (
     <>
@@ -239,42 +353,108 @@ const App = () => {
               </IconButton>
             </Paper>
           </div>
-          <div className={classesTable.root}>
-            <Tabs
-              orientation="vertical"
-              variant="scrollable"
-              value={value}
-              onChange={handleChange}
-              aria-label="Vertical tabs example"
-              className={classesTable.tabs}
-            >
-              <Tab
-                style={{
-                  textTransform:"none"
-                }}
-                label="Contributor Over Time"
-                {...a11yProps(0)}
-              />
-              <Tab
-               style={{
-                textTransform:"none"
+          <div style={{ marginTop: "10px" }}>
+            <CompareComponent
+              list={Object.keys(dataSource)}
+              onDelete={e => {
+                console.log("e : ", e);
+                const newDataSource = omit(dataSource, [e]);
+                setDataSource(newDataSource);
               }}
-                label="Contributor Monthly Activity"
-                {...a11yProps(1)}
-              />
-            </Tabs>
-            <TabPanel value={value} index={0}>
-              <ContirbutorLineChart
-                repoList={contributorRepoList}
-                showAlert={showAlert}
-              />
-            </TabPanel>
-            <TabPanel value={value} index={1}>
-              <ActivityChart
-                repoList={contributorRepoList}
-                showAlert={showAlert}
-              />
-            </TabPanel>
+              onConfirm={e => {
+                if (!e) return;
+                updateChart(e);
+              }}
+            />
+          </div>
+          <div id="chart" style={{ marginTop: "30px" }}>
+            <Tab.Container defaultActiveKey="contributor">
+              <Row>
+                <Col>
+                  <Tab.Content>
+                    <Tab.Pane eventKey="contributor">
+                      <div style={{ marginBottom: "5px" }}>
+                        <ButtonGroup color="secondary" size="small">
+                          <Button
+                            variant={
+                              activeDate === "1month" ? "contained" : "outlined"
+                            }
+                            value="1month"
+                            onClick={e => {
+                              setActiveDate(e.currentTarget.value);
+                            }}
+                          >
+                            1 Month
+                          </Button>
+                          <Button
+                            variant={
+                              activeDate === "3months"
+                                ? "contained"
+                                : "outlined"
+                            }
+                            value="3months"
+                            onClick={e => {
+                              setActiveDate(e.currentTarget.value);
+                            }}
+                          >
+                            3 Months
+                          </Button>
+                          <Button
+                            variant={
+                              activeDate === "6months"
+                                ? "contained"
+                                : "outlined"
+                            }
+                            value="6months"
+                            onClick={e => {
+                              setActiveDate(e.currentTarget.value);
+                            }}
+                          >
+                            6 Months
+                          </Button>
+                          <Button
+                            variant={
+                              activeDate === "1year" ? "contained" : "outlined"
+                            }
+                            value="1year"
+                            onClick={e => {
+                              setActiveDate(e.currentTarget.value);
+                            }}
+                          >
+                            1 Year
+                          </Button>
+                          <Button
+                            variant={
+                              activeDate === "max" ? "contained" : "outlined"
+                            }
+                            value="max"
+                            onClick={e => {
+                              setActiveDate(e.currentTarget.value);
+                            }}
+                          >
+                            Max
+                          </Button>
+                        </ButtonGroup>
+                      </div>
+                      <ReactECharts
+                        option={option}
+                        opts={{ renderer: "svg" }}
+                        ref={e => {
+                          if (e) {
+                            const echartInstance = e.getEchartsInstance();
+                            // then you can use any API of echarts.
+                            window.echartInstance = echartInstance;
+                          }
+                        }}
+                        style={{ height: 700, width: "100%" }}
+                        showLoading={loading}
+                        notMerge
+                      />
+                    </Tab.Pane>
+                  </Tab.Content>
+                </Col>
+              </Row>
+            </Tab.Container>
           </div>
         </div>
       </div>
