+++ src/constants.js
@@ -24,30 +24,3 @@ export const DEFAULT_OPTIONS = {
   },
   series: []
 };
-
-export const DEFAULT_ACTIVITY_OPTIONS = {
-  legend: {
-    top: "5%",
-    data: []
-  },
-  toolbox: {
-    feature: {
-      saveAsImage: {}
-    }
-  },
-  dataset: [],
-  title: {
-    text: "Contributor Monthly Activity"
-  },
-  tooltip: {
-    trigger: "axis"
-  },
-  xAxis: {
-    type: "time",
-    nameLocation: "middle"
-  },
-  yAxis: {
-    name: ""
-  },
-  series: []
-};
